Move merkle tree test into the merkle_tree module
